+++ crates/ntp_usg-client/src/smol_ntp.rs
@@ -27,11 +27,11 @@
 //! # }
 //! ```
 
-use log::debug;
 use smol::net::UdpSocket;
 use std::io;
 use std::net::SocketAddr;
 use std::time::Duration;
+use tracing::debug;
 
 use crate::NtpResult;
 use crate::request::{bind_addr_for, build_request_packet, validate_response};